# Add a specializer to simplify higher-order words.