--- a/src/mirth/elab.mth
+++ b/src/mirth/elab.mth
@@ -314,7 +314,7 @@ def(atoms-turn-last-block-to-arg, Atom List(Atom) -- Atom List(Atom),
             unsnoc dup op match(
                 OP_BLOCK ->
                     { Atom List(Atom) Atom Block }
-                    dip(cod rotl dom!)
+                    dip(dom rotl dom!)
                     { List(Atom) Atom Block }
                     ARG_BLOCK swap args:cons
                     swap,
@@ -325,6 +325,7 @@ def(atoms-turn-last-block-to-arg, Atom List(Atom) -- Atom List(Atom),
 def(ab-op!, Op +AB -- +AB,
     ab-ctx@ >ctx
     ab-token@ >token
+    ab-home@ >home
     dup >op
     elab-op-fresh-sig! dip:>subst
     ab-expand-opsig! >cod >dom
@@ -599,7 +600,8 @@ def(elab-namespace-is-visible, Namespace +AB -- Bool +AB,
     NAMESPACE_ROOT -> F,
     NAMESPACE_PACKAGE -> drop F,
     NAMESPACE_MODULE -> elab-module-is-visible,
-    NAMESPACE_TYCON -> elab-tycon-is-visible)
+    NAMESPACE_TYCON -> elab-tycon-is-visible,
+    NAMESPACE_WORD -> drop F)
 
 def(elab-tycon-is-visible, Tycon +AB -- Bool +AB,
     qname elab-qname-is-visible) # FIXME