--- a/src/mirth/elab.mth
+++ b/src/mirth/elab.mth
@@ -410,6 +410,7 @@ def(elab-op-fresh-sig!, Op -- Subst OpSig,
         OP_EXTERNAL -> type freshen-sig OPSIG_APPLY,
         OP_FIELD -> type freshen-sig OPSIG_APPLY,
         OP_BLOCK -> elab-block-sig!,
+        OP_RUN -> type semifreshen-sig OPSIG_APPLY,
         OP_VAR -> elab-var-sig!,
         OP_MATCH -> elab-match-sig!,
         OP_LAMBDA -> elab-lambda-sig!,
@@ -1010,23 +1011,26 @@ def(create-projectors!, Tag --,
                     T-> pack2
                 )
             ) lbl_lens ~ctx-type !
-            lbl_lens delay(ab-build-word-arrow!(
-                dat is-resource? if(
-                    lbl_get ab-word!
-                    ab-rdip!(
-                        PRIM_CORE_SWAP ab-prim!
-                        PRIM_CORE_RUN ab-prim!
-                    )
-                    lbl_set ab-word!,
+            lbl_lens delay(
+                type dom force-cons?! unwrap unpack2 nip "f" >Name Var.new-auto-run!
+                PARAM L1
+            ) lbl_lens ~params !
 
-                    PRIM_CORE_SWAP ab-prim!
-                    PRIM_CORE_DUP ab-prim!
-                    ab-dip!(
+            lbl_lens delay(ab-build-word-arrow!(
+                lbl_lens params dup ab-lambda!(
+                    /L1 unwrap /PARAM >f
+                    dat is-resource? if(
                         lbl_get ab-word!
-                        PRIM_CORE_SWAP ab-prim!
-                        PRIM_CORE_RUN ab-prim!
+                        ab-rdip!(f> ab-var!)
+                        lbl_set ab-word!,
+
+                        PRIM_CORE_DUP ab-prim!
+                        ab-dip!(
+                            lbl_get ab-word!
+                            f> ab-var!
+                        )
+                        lbl_set ab-word!
                     )
-                    lbl_set ab-word!
                 )
             )) lbl_lens ~arrow !
         )
@@ -1218,7 +1222,7 @@ def(typecheck-everything!, --,
 # TABLE #
 #########
 
-def(TABLE_MAX_COUNT, Nat, 0x10000 >Nat)
+def(TABLE_MAX_COUNT, Nat, 0x20000 >Nat)
 
 def(table-qname, Table Str Int -- QName,
     >arity >Name >name TYCON_TABLE NAMESPACE_TYCON >namespace MKQNAME)