… bin/mirth0.c
… src/mirth/need.mth
… src/std/map.mth
… src/std/maybe.mth
… src/std/prelude.mth
… src/std/str.mth