This PR adds a specializer, which is a compiler pass that converts most uses of higher-order words into first-order words by fixing their arguments. This works pretty well in mirth because higher-order words are defined via parameters, in a way that interacts little with the stack. So the specializer really only has to expand definitions (carefully, to preserve types and because of recursion). (But this also means that if you *are* passing around closures manually on the stack, the specializer won't help very much.)

The biggest advantage of doing this expansion is that it should be a lot easier to optimize with mostly first-order functions. However this step also increases the size of the compiler output. Although the mirth compiler got faster, the compiler dev iteration time (the time it takes to run `make`) remains the same on my machine:

- `time make` went from 37s to 37s
- `time make bin/mirth1.c` went from 4.9s to 4.2s